refactor: use single script to apply/revoke rules and permissions, use install script
